xxw/fraud_api/requirements.txt
flask==1.0.2
marshmallow==2.15.3
marshmallow-mongoengine==0.9.1
requests==2.18.4
kazoo==2.5.0
raven==6.9.0
blinker==1.4
pytest==3.6.1
pytest-cov==2.5.1
pytest-flake8===1.0.1
pyinstaller==3.3.1
isort==4.3.4
black==18.6b4
redis==2.10.6
flask-restful==0.3.6
meinheld==0.6.1
gunicorn==19.8.1
flask_marshmallow
schema==0.6.8
flask-mongoengine==0.9.5
